Split new_functions into topics; improved error handling; polishing
