+++ pyproject.toml
@@ -1,4 +1,5 @@
 [tool.poetry]
+package-mode = false
 name = "nmrlipids"
 version = "0.1.0"
 description = "Tutorial for accessing NMR lipids databank through their API and training a neural network"
@@ -14,7 +15,7 @@ matplotlib = "^3.9.2"
 tqdm = "^4.67.0"
 pyyaml = "^6.0.2"
 mdanalysis = "^2.7.0"
-scikit-learn = ">=1.5.2,<1.5.3"
+scikit-learn = "1.5.2"
 ipywidgets = "^8.1.5"
 pathlib = "^1.0.1"
 scikeras = "^0.13.0"
